+++ src/app/constants.rs
@@ -1,0 +1,7 @@
+pub(crate) const DEFAULT_COLS: u16 = 100;
+pub(crate) const DEFAULT_ROWS: u16 = 30;
+pub(crate) const SIDEBAR_WIDTH: f32 = 306.0;
+pub(crate) const TAB_BAR_HEIGHT: f32 = 52.0;
+pub(crate) const TERMINAL_PADDING_X: f32 = 32.0;
+pub(crate) const TERMINAL_PADDING_Y: f32 = 32.0;
+pub(crate) const TERMINAL_KEY_CONTEXT: &str = "AshellTerminal";
