+++ src/app/startup.rs
@@ -1,0 +1,166 @@
+use gpui::{App, AppContext as _, Bounds, WindowOptions, point, px, size};
+use gpui_component::Root;
+
+use crate::session::config::ConfigStore;
+use crate::Ashell;
+
+#[cfg(target_os = "macos")]
+pub(crate) fn sync_macos_launch_environment() {
+    let shell = std::env::var("SHELL").unwrap_or_else(|_| "/bin/zsh".to_string());
+    let Ok(output) = std::process::Command::new(&shell).args(["-l", "-c", "env -0"]).output() else {
+        return;
+    };
+    if !output.status.success() {
+        return;
+    }
+
+    for entry in output.stdout.split(|b| *b == 0) {
+        if entry.is_empty() {
+            continue;
+        }
+        let Some(eq) = entry.iter().position(|b| *b == b'=') else {
+            continue;
+        };
+        let Ok(key) = std::str::from_utf8(&entry[..eq]) else {
+            continue;
+        };
+        let Ok(value) = std::str::from_utf8(&entry[eq + 1..]) else {
+            continue;
+        };
+
+        let should_import = matches!(
+            key,
+            "PATH"
+                | "MANPATH"
+                | "INFOPATH"
+                | "LANG"
+                | "LC_ALL"
+                | "LC_CTYPE"
+                | "SHELL"
+                | "HOME"
+                | "HOMEBREW_PREFIX"
+                | "HOMEBREW_CELLAR"
+                | "HOMEBREW_REPOSITORY"
+        ) || key.starts_with("LC_");
+
+        if should_import {
+            unsafe {
+                std::env::set_var(key, value);
+            }
+        }
+    }
+}
+
+#[cfg(not(target_os = "macos"))]
+pub(crate) fn sync_macos_launch_environment() {}
+
+pub(crate) fn open_main_window(cx: &mut App) {
+    let mut window_options = WindowOptions::default();
+
+    #[cfg(not(target_os = "macos"))]
+    if let Ok(img) = image::load_from_memory(include_bytes!("../../assets/icons/ashell.png")) {
+        window_options.icon = Some(std::sync::Arc::new(img.into_rgba8()));
+    }
+
+    let config = ConfigStore::load().unwrap_or_else(|_| ConfigStore::in_memory());
+    if let Some(bounds) = config.window_bounds() {
+        window_options.window_bounds = Some(match bounds {
+            crate::session::config::SavedWindowBounds::Fullscreen {
+                x,
+                y,
+                width,
+                height,
+            } => gpui::WindowBounds::Fullscreen(Bounds::new(
+                point(px(*x), px(*y)),
+                size(px(*width), px(*height)),
+            )),
+            crate::session::config::SavedWindowBounds::Maximized {
+                x,
+                y,
+                width,
+                height,
+            } => gpui::WindowBounds::Maximized(Bounds::new(
+                point(px(*x), px(*y)),
+                size(px(*width), px(*height)),
+            )),
+            crate::session::config::SavedWindowBounds::Windowed {
+                x,
+                y,
+                width,
+                height,
+            } => gpui::WindowBounds::Windowed(Bounds::new(
+                point(px(*x), px(*y)),
+                size(px(*width), px(*height)),
+            )),
+        });
+    } else if let Some(display) = cx.displays().first().cloned() {
+        let display_bounds = display.bounds();
+        let width = display_bounds.size.width * 0.8;
+        let height = display_bounds.size.height * 0.9;
+
+        let x = display_bounds.origin.x + (display_bounds.size.width - width) / 2.0;
+
+        #[cfg(target_os = "macos")]
+        let y = display_bounds.origin.y;
+        #[cfg(not(target_os = "macos"))]
+        let y = display_bounds.origin.y + (display_bounds.size.height - height) / 2.0;
+
+        window_options.window_bounds = Some(gpui::WindowBounds::Windowed(Bounds::new(
+            point(x, y),
+            size(width, height),
+        )));
+    }
+
+    cx.open_window(window_options, |window, cx| {
+        window.activate_window();
+        window.set_window_title("ashell");
+        gpui_component::Theme::sync_system_appearance(Some(window), cx);
+        let view = cx.new(|cx| Ashell::new(window, cx));
+
+        let workspace_panels_clone = view.read(cx).workspace_panels.clone();
+        let body_panels_clone = view.read(cx).body_panels.clone();
+        window.on_window_should_close(cx, move |window: &mut gpui::Window, cx: &mut gpui::App| {
+            let mut config = ConfigStore::load().unwrap_or_else(|_| ConfigStore::in_memory());
+            let current_bounds = window.window_bounds();
+            let saved_bounds = match current_bounds {
+                gpui::WindowBounds::Fullscreen(b) => crate::session::config::SavedWindowBounds::Fullscreen {
+                    x: b.origin.x.into(),
+                    y: b.origin.y.into(),
+                    width: b.size.width.into(),
+                    height: b.size.height.into(),
+                },
+                gpui::WindowBounds::Maximized(b) => crate::session::config::SavedWindowBounds::Maximized {
+                    x: b.origin.x.into(),
+                    y: b.origin.y.into(),
+                    width: b.size.width.into(),
+                    height: b.size.height.into(),
+                },
+                gpui::WindowBounds::Windowed(b) => crate::session::config::SavedWindowBounds::Windowed {
+                    x: b.origin.x.into(),
+                    y: b.origin.y.into(),
+                    width: b.size.width.into(),
+                    height: b.size.height.into(),
+                },
+            };
+            let workspace_sizes: Vec<f32> = workspace_panels_clone
+                .read(cx)
+                .sizes()
+                .iter()
+                .map(|s| s.into())
+                .collect();
+            let body_sizes: Vec<f32> = body_panels_clone
+                .read(cx)
+                .sizes()
+                .iter()
+                .map(|s| s.into())
+                .collect();
+            config.set_layout_state(Some(saved_bounds), Some(workspace_sizes), Some(body_sizes));
+            let _ = config.save();
+            true
+        });
+
+        cx.new(|cx| Root::new(view, window, cx))
+    })
+    .expect("failed to open window");
+}
+
